Mangrove Trees/SawGrass
When first loading in the folder, it will state that a material uses the texture as a normal map and that the normal map needs to be marked as one. Simply click on fix now and the issue will resolve itself. To bring the objects into the scene, drag the .st file (outline of tree/grass) into the scene, and the LOD and settings should be implemented instantly. From there it can be resized and rotated accordingly.