Set-up Tree and Sawgrass Interactables
Set up the trees and the sawgrass to be interactable objects that pop up specific information when triggered. When setting up the rest of the interactable items, simply duplicate the pop-up item that is attached to the main camera for the canoe, and change the title and the text in the scroll view drop-down. When adding the script component to make the item interactable you need to make sure to create a new interactable layer and set the object to that layer, and also change that layer in the script, otherwise, it will trigger the rest of the pop-ups. The box and canvas object placed originally are still there, but they are both turned off because interacting with them since the box is not in its layer causes issues with the rest of the interactable objects.